// Angular/crudPacientes/src/app/modelos/usuario.interface.ts
export interface PacienteI {
    PacienteId:      number;
    DNI:             string;
    Nombre:          string;
    Direccion:       string;
    CodigoPostal:    string;
    Telefono:        string;
    Genero:          string;
    FechaNacimiento: string;
    Correo:          string;
    edit:            boolean;
    data:            boolean,
}